fix: select contents on next frame
With DevTools it looks like when you select text right-click menu is triggered automatically.
This happens on all websites, we can prevent it on initial click by delaying it
to next frame, but if you click again it will be triggered anyway.

Not sure if this is worth keeping, it doesn't cause real-world issues AFAIK.